+++ docs/sample_code.lua
@@ -17,7 +17,7 @@ end
 
 function execute_code()
 
-    check()
+    _check()
 
 	participant_result = {}
 
@@ -31,6 +31,8 @@ function execute_code()
 
 	for i = lower,upper,1
 	do
+	    if i % 100000 == 0 then _progress(i) end
+
 		if (prime % i == 0 )
 		then
 			participant_result.divisor = i